Restructure agents into staff member directory
