+++ lib/handlers/handleTeamsMessages.js
@@ -8,9 +8,9 @@ import {
 } from "../utils/adaptiveCards.js";
 import { getClientDirectories } from "../utils/fileStorageAndRetrieval.js";
 import { createTeamsUpdate } from "../utils/utils.js";
-import { runProcessing } from "../../agents/processingAgent.js";
-import { runAuditorNotes } from "../../agents/auditorNotesAgent.js";
-import { callHeadAgent } from "../../agents/headAgent.js";
+import { runProcessing } from "../../staff members/agents/processingAgent.js";
+import { runAuditorNotes } from "../../staff members/agents/auditorNotesAgent.js";
+import { callHeadAgent } from "../../staff members/agents/headAgent.js";
 /**
  * Processes incoming Teams bot messages and routes them to appropriate handlers.
  *
